Fixed some issues in the plotting
Accepted input coordinate types seem to have changed in either Plots.jl or some dependency of it.
Earlier it accepted an array of points to contain both x and y values. Now it only works with two separate vectors for x and y values, so I needed to write the plotting a little more carefully.